+++ Linkee-api/src/main/java/com/linkee/linkeeapi/report/command/application/dto/request/UpdateReportActionRequestDto.java
@@ -1,4 +1,4 @@
-package com.linkee.linkeeapi.report.model.dto.request;
+package com.linkee.linkeeapi.report.command.application.dto.request;
 
 import lombok.*;
