+++ Linkee-api/src/main/java/com/linkee/linkeeapi/report/command/infrastructure/repository/ReportRepository.java
@@ -1,0 +1,8 @@
+package com.linkee.linkeeapi.report.command.infrastructure.repository;
+
+
+import com.linkee.linkeeapi.report.command.domain.aggregate.Report;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ReportRepository extends JpaRepository<Report, Long> {
+}
