--- Linkee-api/src/main/java/com/linkee/linkeeapi/report/repository/ReportRepository.java
@@ -1,7 +1,0 @@
-package com.linkee.linkeeapi.report.repository;
-
-import com.linkee.linkeeapi.report.model.entity.Report;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ReportRepository extends JpaRepository<Report, Long> {
-}
